create wdio-utils package to move common utils in there
