+++ packages/wdio-cli/package.json
@@ -38,6 +38,7 @@
     "@wdio/config": "^5.4.18",
     "@wdio/interface": "^5.4.6",
     "@wdio/logger": "^5.4.6",
+    "@wdio/utils": "^5.4.18",
     "async-exit-hook": "^2.0.1",
     "chalk": "^2.3.2",
     "chokidar": "^2.0.4",
