+++ packages/wdio-cli/tests/utils.test.js
@@ -1,26 +1,4 @@
-import logger from '@wdio/logger'
-
-import { filterPackageName, getLauncher, runServiceHook, getRunnerName } from '../src/utils'
-
-jest.mock('@wdio/config', () => {
-    class LauncherMock {
-        constructor () {
-            this.onPrepare = jest.fn()
-            this.onComplete = jest.fn()
-        }
-    }
-
-    return {
-        initialisePlugin: jest.fn()
-            .mockImplementationOnce(
-                () => LauncherMock)
-            .mockImplementationOnce(
-                () => undefined)
-            .mockImplementationOnce(
-                () => { throw new Error('Couldn\'t find plugin') })
-            .mockImplementationOnce(() => { throw new Error('buhh') })
-    }
-})
+import { filterPackageName, runServiceHook, getRunnerName } from '../src/utils'
 
 test('filterPackageName', () => {
     const reporter = [
@@ -35,50 +13,6 @@ test('filterPackageName', () => {
         '@wdio/junit-reporter',
         'wdio-random-reporter'
     ])
-})
-
-test('getLauncher', () => {
-    expect(getLauncher({})).toEqual([])
-
-    const inlineService = {
-        onPrepare: jest.fn(),
-        beforeTest: jest.fn()
-    }
-    expect(getLauncher({
-        services: [
-            inlineService,
-            'unscoped',
-            'scoped',
-            'non-existing'
-        ]
-    })).toHaveLength(2)
-})
-
-test('getLauncher not failing on syntax error', () => {
-    expect(() => getLauncher({
-        services: ['other-unscoped']
-    })).toHaveLength(0)
-    expect(logger().error).toBeCalledTimes(1)
-})
-
-test('getLauncher sets correct service scope', () => {
-    const hookSuccess = jest.fn()
-
-    const inlineService = {
-        onPrepare() {
-            this._otherMethod()
-        },
-        _otherMethod: hookSuccess
-    }
-
-    const launcher = getLauncher({
-        services: [
-            inlineService
-        ]
-    })
-
-    runServiceHook(launcher, 'onPrepare', 1, true, 'abc')
-    expect(hookSuccess).toBeCalledTimes(1)
 })
 
 test('runServiceHook', () => {
