+++ packages/wdio-config/src/index.js
@@ -1,5 +1,5 @@
 import ConfigParser from './lib/ConfigParser'
-import { validateConfig, getSauceEndpoint, detectBackend, initialisePlugin } from './utils'
+import { validateConfig, getSauceEndpoint, detectBackend } from './utils'
 import {
     wrapCommand, runFnInFiberContext, runTestInFiberContext, executeHooksWithArgs,
     hasWdioSyncSupport
@@ -10,7 +10,6 @@ export {
     validateConfig,
     getSauceEndpoint,
     detectBackend,
-    initialisePlugin,
     ConfigParser,
 
     /**
