+++ packages/wdio-utils/.npmignore
@@ -1,0 +1,2 @@
+src
+tests
