+++ packages/wdio-utils/.npmrc
@@ -1,0 +1,1 @@
+message = "@wdio/utils: bump version %s"
