+++ packages/wdio-utils/README.md
@@ -1,0 +1,4 @@
+WDIO Repl
+=========
+
+> A WDIO helper utility to provide a repl interface WebdriverIO
