+++ packages/wdio-utils/package.json
@@ -1,9 +1,9 @@
 {
-  "name": "@wdio/runner",
-  "version": "5.4.19",
-  "description": "A WebdriverIO service that runs tests in arbitrary environments",
+  "name": "@wdio/utils",
+  "version": "5.4.18",
+  "description": "A WDIO helper utility to provide several utility function used across the project.",
   "author": "Christian Bromann <christian@saucelabs.com>",
-  "homepage": "https://github.com/webdriverio/webdriverio/tree/master/packages/wdio-runner",
+  "homepage": "https://github.com/webdriverio/webdriverio/tree/master/packages/wdio-utils",
   "license": "MIT",
   "main": "./build/index",
   "engines": {
@@ -22,20 +22,17 @@
     "url": "git://github.com/webdriverio/webdriverio.git"
   },
   "keywords": [
+    "webdriverio",
     "webdriver",
     "wdio",
-    "wdio-reporter"
+    "utils"
   ],
   "bugs": {
     "url": "https://github.com/webdriverio/webdriverio/issues"
   },
   "dependencies": {
-    "@wdio/config": "^5.4.18",
     "@wdio/logger": "^5.4.6",
-    "@wdio/utils": "^5.4.18",
-    "deepmerge": "^2.0.1",
-    "gaze": "^1.1.2",
-    "webdriverio": "^5.4.19"
+    "deepmerge": "^3.2.0"
   },
   "publishConfig": {
     "access": "public"
