+++ packages/wdio-utils/src/index.js
@@ -1,0 +1,7 @@
+import initialisePlugin from './initialisePlugin'
+import initialiseServices from './initialiseServices'
+
+export {
+    initialisePlugin,
+    initialiseServices
+}
